+++ bundles/mapping/mapmodule/mapmodule.ol3.js
@@ -83,7 +83,7 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
         _getMapCenter: function() {
             return this._map.getView().getCenter();
         },
-        _getMapZoom: function() {
+        getMapZoom: function() {
             return this._map.getView().getZoom();
         },
         _getMapLayersByName: function(layerName) {
@@ -342,11 +342,6 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
             }
 
         },
-
-        getZoomLevel: function() {
-            return this._map.getView().getZoom();
-        },
-
         /**
          * @method _updateDomain
          * @private
@@ -362,7 +357,7 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
             var sandbox = this._sandbox;
             var mapVO = sandbox.getMap();
             var lonlat = this._getMapCenter();
-            var zoom = this._getMapZoom();
+            var zoom = this.getMapZoom();
             mapVO.moveTo(lonlat[0], lonlat[1], zoom);
 
             mapVO.setScale(this.getMapScale());
@@ -458,33 +453,24 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
 
         /**
          * @method transformCoordinates
-         * Deprecated
          * Transforms coordinates from given projection to the maps projectino.
-         * @param {OpenLayers.LonLat} pLonlat
+         * @param {Object} pLonlat object with lon and lat keys
          * @param {String} srs projection for given lonlat params like "EPSG:4326"
-         * @return {OpenLayers.LonLat} transformed coordinates
+         * @return {Object} transformed coordinates as object with lon and lat keys
          */
         transformCoordinates: function (pLonlat, srs) {
-            this.getSandbox().printWarn(
-                'transformCoordinates is deprecated. Use _transformCoordinates instead if called from plugin. Otherwise, use Requests instead.'
-            );
-
-            return this._transformCoordinates(pLonlat, srs);
+            if(!srs || this.getProjection() === srs) {
+                return pLonlat;
+            }
+            // TODO: check that srs definition exists as in OL2
+            //var transformed = new ol.proj.fromLonLat([pLonlat.lon, pLonlat.lat], this.getProjection());
+            var transformed = ol.proj.transform([pLonlat.lon, pLonlat.lat], srs, this.getProjection());
+            return {
+              lon : transformed[0],
+              lat : transformed[1]
+            };
         },
 
-        /**
-         * @method _transformCoordinates
-         * Transforms coordinates from given projection to the maps projectino.
-         * @param {OpenLayers.LonLat} pLonlat
-         * @param {String} srs projection for given lonlat params like "EPSG:4326"
-         * @return {OpenLayers.LonLat} transformed coordinates
-         */
-        _transformCoordinates: function (pLonlat, srs) {
-            return pLonlat.transform(
-                new OpenLayers.Projection(srs),
-                this.getMap().getProjectionObject()
-            );
-        },
         /**
          * @property eventHandlers
          * @static
@@ -677,7 +663,7 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
          */
         setZoomLevel: function (newZoomLevel, suppressEvent) {
             if (newZoomLevel < 0 || newZoomLevel > this.getMaxZoomLevel()) {
-                newZoomLevel = this._getMapZoom();
+                newZoomLevel = this.getMapZoom();
             }
             this._map.getView().setZoom(newZoomLevel);
             /*
@@ -757,28 +743,39 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
          * Moves the map to the given position.
          * NOTE! Doesn't send an event if zoom level is not changed.
          * Call notifyMoveEnd() afterwards to notify other components about changed state.
-         * @param {OpenLayers.LonLat} lonlat coordinates to move the map to
+         * @param {Number[] | Object} lonlat coordinates to move the map to
          * @param {Number} zoomAdjust relative change to the zoom level f.ex -1 (optional)
          * @param {Boolean} pIsDragging true if the user is dragging the map to a new location currently (optional)
          */
         moveMapToLonLat: function (lonlat, zoomAdjust, pIsDragging) {
-            // TODO: openlayers has isValidLonLat(); maybe use it here
-            var isDragging = (pIsDragging === true);
-            // TODO check if panTo (still) breaks IE9+ and if not, should we use it
-            // using panTo BREAKS IE on startup so do not
+            // parse lonlat if it's given as an array instead of {lon : x, lat : y}
+            lonlat = this.normalizeLonLat(lonlat);
             // should we spam events on dragmoves?
-            this._map.getView().setCenter([lonlat[0], lonlat[1]]);
+            this._map.getView().setCenter([lonlat.lon, lonlat.lat]);
 
             if (zoomAdjust) {
                 this.adjustZoomLevel(zoomAdjust, true);
             }
             this._updateDomainImpl();
         },
-
-        setMapCenter: function (lonlat, zoom) {
+        /**
+         * @method centerMap
+         * Moves the map to the given position and zoomlevel.
+         * @param {Number[] | Object} lonlat coordinates to move the map to
+         * @param {Number} zoomLevel absolute zoomlevel to set the map to
+         * @param {Boolean} suppressEnd true to NOT send an event about the map move
+         *  (other components wont know that the map has moved, only use when chaining moves and
+         *     wanting to notify at end of the chain for performance reasons or similar) (optional)
+         */
+        centerMap: function (lonlat, zoom, suppressEnd) {
+            // TODO: we have isValidLonLat(); maybe use it here
+            lonlat = this.normalizeLonLat(lonlat);
             this._map.getView().setCenter([lonlat[0], lonlat[1]]);
             this._map.getView().setZoom(zoom);
             this._updateDomainImpl();
+            if (suppressEnd !== true) {
+                this.notifyMoveEnd();
+            }
         },
 
         _getMapCenter: function () {
@@ -862,8 +859,22 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
 
             var lonlat = this._getMapCenter();
             this._updateDomainImpl();
-            var evt = sandbox.getEventBuilder('AfterMapMoveEvent')(lonlat.lon, lonlat.lat, this._getMapZoom(), false, this.getMapScale(), creator);
+            var evt = sandbox.getEventBuilder('AfterMapMoveEvent')(lonlat.lon, lonlat.lat, this.getMapZoom(), false, this.getMapScale(), creator);
             sandbox.notifyAll(evt);
+        },
+        /**
+         * Get map max extent.
+         * @method getMaxExtent
+         * @return {Object} max extent
+         */
+        getMaxExtent: function(){
+            var bbox = this.getSandbox().getMap().getBbox();
+            return {
+                bottom: bbox.bottom,
+                left: bbox.left,
+                right: bbox.right,
+                top: bbox.top
+            };
         }
 
     }, {